main/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
groq==0.13.1
ipykernel==6.29.5
matplotlib==3.10.0
numpy==1.26.4
pillow==11.0.0
python-dotenv==1.0.1
tensorflow==2.18.0
scikit-learn==1.6.1